requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cssselect2==0.8.0
Django==5.2.3
django-environ==0.12.0
django-filter==25.1
django-modeltranslation==0.19.15
djangorestframework==3.16.0
djangorestframework-csv==3.0.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
easyocr==1.7.2
filelock==3.18.0
fonttools==4.58.4
fsspec==2025.5.1
gunicorn==23.0.0
idna==3.10
imageio==2.37.0
inflection==0.5.1
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
ninja==1.11.1.4
numpy==2.3.0
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pyclipper==1.3.0.post6
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.17.2
python-bidi==0.6.6
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.2
requests==2.32.4
scikit-image==0.25.2
scipy==1.15.3
setuptools==80.9.0
shapely==2.1.1
sqlparse==0.5.3
sympy==1.14.0
tifffile==2025.6.11
tinycss2==1.4.0
tinyhtml5==2.0.0
torch==2.7.1
torchvision==0.22.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
weasyprint==65.1
webencodings==0.5.1
whitenoise==6.9.0
zopfli==0.2.3.post1
